requirements_merged.txt
#front
python-dotenv
sounddevice
pyaudio
soundfile
streamlit
ipykernel
ipywidgets
jupyter

# back
fastapi 
uvicorn
transformers
pytube
gtts
huggingface
ipykernel
ipywidgets
jupyter
openai-whisper
pydub
tqdm
pyaudio

#+
accelerate